# feature: stock 조회 API 완성

+++ src/main/java/com/minsoo/co/tireerpserver/model/dto/stock/TireStockResponse.java
@@ -14,8 +14,8 @@ public class TireStockResponse {
     private Long averageOfPurchasePrice;
 
     @JsonProperty("sum_of_stock")
-    private Integer sumOfStock;
+    private Long sumOfStock;
 
     @JsonProperty("number_of_dot")
-    private Integer numberOfDot;
+    private Long numberOfDot;
 }
